refactor(selectable): Convert into typescript
